stub privilege checking + delaying message processing
This introduces a fake method for privilege checking which simply
always denies access. The bigger change is in the handling of the new
"later" result of policy checking: this result must be rolled up to
the caller who then needs to decide what to do with the message.

Messages in the connection->incoming_messages list get put back into
the list. The intention is to disable dispatching of that connection
until the policy check on that message can succeed; this still needs
to be implemented.

Many other call paths will simply fail when policy checking does not
work right away.