activation: handle BUS_RESULT_LATER
This initial shot at handling a message which caused a service to be
activated and then cannot be delivered immediately because the policy
check takes time is good enough if the sender sends only one message.
It does not yet handle the situation correctly where the sender sends
multiple messages.

It is worth pointing out that the service gets auto-activated by
dbus-daemon without first checking whether the sender is allowed to
talk to the service. This is a weakness that also applies to the
traditional <allow/deny> rules.

Perhaps we should add a _can_send check before initiating the
auto-activation.