Update to dbus-1.10.6
Note: from version 1.10 it's required to have systemd support to be enabled not
only in dbus-daemon code but also in libdbus.  This is because maintainers removed
built-in systemd support code (dbus/sd-daemon.c) and now depend on libsystemd-
provided functionality.

Additionally, update pkgconfig name to libsystemd (as required by configure.ac)

Change-Id: I2056086a9281543695643ac31fd3ab8648a4d205